opt_merge: factor out hashing code across incremental and parallel
