+++ passes/opt/opt_merge.cc
@@ -24,6 +24,7 @@
 #include "kernel/celltypes.h"
 #include "kernel/threading.h"
 #include "libs/sha1/sha1.h"
+#include "passes/opt/opt_merge_common.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <algorithm>
@@ -35,8 +36,7 @@
 USING_YOSYS_NAMESPACE
 PRIVATE_NAMESPACE_BEGIN
 
-template <typename T, typename U>
-inline Hasher hash_pair(const T &t, const U &u) { return hash_ops<std::pair<T, U>>::hash(t, u); }
+using OptMergeCommon::CellHasher;
 
 // Some cell and its hash value.
 struct CellHash
@@ -96,178 +96,19 @@ struct FoundDuplicates
 	std::vector<DuplicateCell> duplicates;
 };
 
-struct OptMergeThreadWorker
+struct OptMergeThreadWorker : public CellHasher
 {
 	const RTLIL::Module *module;
-	const SigMap &assign_map;
-	const FfInitVals &initvals;
 	const CellTypes &ct;
 	int workers;
 	bool mode_share_all;
 	bool mode_keepdc;
 
-	static Hasher hash_pmux_in(const SigSpec& sig_s, const SigSpec& sig_b, Hasher h)
-	{
-		int s_width = GetSize(sig_s);
-		int width = GetSize(sig_b) / s_width;
-
-		hashlib::commutative_hash comm;
-		for (int i = 0; i < s_width; i++)
-			comm.eat(hash_pair(sig_s[i], sig_b.extract(i*width, width)));
-
-		return comm.hash_into(h);
-	}
-
-	static void sort_pmux_conn(dict<RTLIL::IdString, RTLIL::SigSpec> &conn)
-	{
-		const SigSpec &sig_s = conn.at(ID::S);
-		const SigSpec &sig_b = conn.at(ID::B);
-
-		int s_width = GetSize(sig_s);
-		int width = GetSize(sig_b) / s_width;
-
-		vector<pair<SigBit, SigSpec>> sb_pairs;
-		for (int i = 0; i < s_width; i++)
-			sb_pairs.push_back(pair<SigBit, SigSpec>(sig_s[i], sig_b.extract(i*width, width)));
-
-		std::sort(sb_pairs.begin(), sb_pairs.end());
-
-		conn[ID::S] = SigSpec();
-		conn[ID::B] = SigSpec();
-
-		for (auto &it : sb_pairs) {
-			conn[ID::S].append(it.first);
-			conn[ID::B].append(it.second);
-		}
-	}
-
-	Hasher hash_cell_inputs(const RTLIL::Cell *cell, Hasher h) const
-	{
-		// TODO: when implemented, use celltypes to match:
-		// (builtin || stdcell) && (unary || binary) && symmetrical
-		if (cell->type.in(ID($and), ID($or), ID($xor), ID($xnor), ID($add), ID($mul),
-				ID($logic_and), ID($logic_or), ID($_AND_), ID($_OR_), ID($_XOR_))) {
-			hashlib::commutative_hash comm;
-			comm.eat(assign_map(cell->getPort(ID::A)));
-			comm.eat(assign_map(cell->getPort(ID::B)));
-			h = comm.hash_into(h);
-		} else if (cell->type.in(ID($reduce_xor), ID($reduce_xnor))) {
-			SigSpec a = assign_map(cell->getPort(ID::A));
-			a.sort();
-			h = a.hash_into(h);
-		} else if (cell->type.in(ID($reduce_and), ID($reduce_or), ID($reduce_bool))) {
-			SigSpec a = assign_map(cell->getPort(ID::A));
-			a.sort_and_unify();
-			h = a.hash_into(h);
-		} else if (cell->type == ID($pmux)) {
-			SigSpec sig_s = assign_map(cell->getPort(ID::S));
-			SigSpec sig_b = assign_map(cell->getPort(ID::B));
-			h = hash_pmux_in(sig_s, sig_b, h);
-			h = assign_map(cell->getPort(ID::A)).hash_into(h);
-		} else {
-			hashlib::commutative_hash comm;
-			for (const auto& [port, sig] : cell->connections()) {
-				if (cell->output(port))
-					continue;
-				comm.eat(hash_pair(port, assign_map(sig)));
-			}
-			h = comm.hash_into(h);
-			if (cell->is_builtin_ff())
-				h = initvals(cell->getPort(ID::Q)).hash_into(h);
-		}
-		return h;
-	}
-
-	static Hasher hash_cell_parameters(const RTLIL::Cell *cell, Hasher h)
-	{
-		hashlib::commutative_hash comm;
-		for (const auto& param : cell->parameters) {
-			comm.eat(param);
-		}
-		return comm.hash_into(h);
-	}
-
-	Hasher hash_cell_function(const RTLIL::Cell *cell, Hasher h) const
-	{
-		h.eat(cell->type);
-		h = hash_cell_inputs(cell, h);
-		h = hash_cell_parameters(cell, h);
-		return h;
-	}
-
-	bool compare_cell_parameters_and_connections(const RTLIL::Cell *cell1, const RTLIL::Cell *cell2) const
-	{
-		if (cell1 == cell2) return true;
-		if (cell1->type != cell2->type) return false;
-
-		if (cell1->parameters != cell2->parameters)
-			return false;
-		if (cell1->connections_.size() != cell2->connections_.size())
-			return false;
-		for (const auto &it : cell1->connections_)
-			if (!cell2->connections_.count(it.first))
-				return false;
-
-		decltype(Cell::connections_) conn1, conn2;
-		conn1.reserve(cell1->connections_.size());
-		conn2.reserve(cell1->connections_.size());
-
-		for (const auto &it : cell1->connections_) {
-			if (cell1->output(it.first)) {
-				if (it.first == ID::Q && cell1->is_builtin_ff()) {
-					// For the 'Q' output of state elements,
-					//   use the (* init *) attribute value
-					conn1[it.first] = initvals(it.second);
-					conn2[it.first] = initvals(cell2->getPort(it.first));
-				}
-				else {
-					conn1[it.first] = RTLIL::SigSpec();
-					conn2[it.first] = RTLIL::SigSpec();
-				}
-			}
-			else {
-				conn1[it.first] = assign_map(it.second);
-				conn2[it.first] = assign_map(cell2->getPort(it.first));
-			}
-		}
-
-		if (cell1->type.in(ID($and), ID($or), ID($xor), ID($xnor), ID($add), ID($mul),
-				ID($logic_and), ID($logic_or), ID($_AND_), ID($_OR_), ID($_XOR_))) {
-			if (conn1.at(ID::A) < conn1.at(ID::B)) {
-				std::swap(conn1[ID::A], conn1[ID::B]);
-			}
-			if (conn2.at(ID::A) < conn2.at(ID::B)) {
-				std::swap(conn2[ID::A], conn2[ID::B]);
-			}
-		} else
-		if (cell1->type.in(ID($reduce_xor), ID($reduce_xnor))) {
-			conn1[ID::A].sort();
-			conn2[ID::A].sort();
-		} else
-		if (cell1->type.in(ID($reduce_and), ID($reduce_or), ID($reduce_bool))) {
-			conn1[ID::A].sort_and_unify();
-			conn2[ID::A].sort_and_unify();
-		} else
-		if (cell1->type == ID($pmux)) {
-			sort_pmux_conn(conn1);
-			sort_pmux_conn(conn2);
-		}
-
-		return conn1 == conn2;
-	}
-
-	bool has_dont_care_initval(const RTLIL::Cell *cell) const
-	{
-		if (!cell->is_builtin_ff())
-			return false;
-
-		return !initvals(cell->getPort(ID::Q)).is_fully_def();
-	}
-
 	OptMergeThreadWorker(const RTLIL::Module *module, const FfInitVals &initvals,
 			const SigMap &assign_map, const CellTypes &ct, int workers,
 			bool mode_share_all, bool mode_keepdc) :
-		module(module), assign_map(assign_map), initvals(initvals), ct(ct),
+		CellHasher(assign_map, initvals, /*apply_sigmap=*/true),
+		module(module), ct(ct),
 		workers(workers), mode_share_all(mode_share_all), mode_keepdc(mode_keepdc)
 	{
 	}
