+++ passes/opt/opt_merge_inc.cc
@@ -23,6 +23,7 @@
 #include "kernel/log.h"
 #include "kernel/celltypes.h"
 #include "libs/sha1/sha1.h"
+#include "passes/opt/opt_merge_common.h"
 #include <stdlib.h>
 #include <stdio.h>
 #include <algorithm>
@@ -34,8 +35,7 @@
 USING_YOSYS_NAMESPACE
 PRIVATE_NAMESPACE_BEGIN
 
-template <typename T, typename U>
-inline Hasher hash_pair(const T &t, const U &u) { return hash_ops<std::pair<T, U>>::hash(t, u); }
+using OptMergeCommon::CellHasher;
 
 struct OptMergeIncWorker
 {
@@ -47,168 +47,11 @@ struct OptMergeIncWorker
 
 	CellTypes ct;
 	int total_count;
-
-	static Hasher hash_pmux_in(const SigSpec& sig_s, const SigSpec& sig_b, Hasher h)
-	{
-		int s_width = GetSize(sig_s);
-		int width = GetSize(sig_b) / s_width;
-
-		hashlib::commutative_hash comm;
-		for (int i = 0; i < s_width; i++)
-			comm.eat(hash_pair(sig_s[i], sig_b.extract(i*width, width)));
-
-		return comm.hash_into(h);
-	}
-
-	static void sort_pmux_conn(dict<RTLIL::IdString, RTLIL::SigSpec> &conn)
-	{
-		SigSpec sig_s = conn.at(ID::S);
-		SigSpec sig_b = conn.at(ID::B);
-
-		int s_width = GetSize(sig_s);
-		int width = GetSize(sig_b) / s_width;
-
-		vector<pair<SigBit, SigSpec>> sb_pairs;
-		for (int i = 0; i < s_width; i++)
-			sb_pairs.push_back(pair<SigBit, SigSpec>(sig_s[i], sig_b.extract(i*width, width)));
-
-		std::sort(sb_pairs.begin(), sb_pairs.end());
-
-		conn[ID::S] = SigSpec();
-		conn[ID::B] = SigSpec();
-
-		for (auto &it : sb_pairs) {
-			conn[ID::S].append(it.first);
-			conn[ID::B].append(it.second);
-		}
-	}
-
-	Hasher hash_cell_inputs(const RTLIL::Cell *cell, Hasher h) const
-	{
-		// TODO: when implemented, use celltypes to match:
-		// (builtin || stdcell) && (unary || binary) && symmetrical
-		if (cell->type.in(ID($and), ID($or), ID($xor), ID($xnor), ID($add), ID($mul),
-				ID($logic_and), ID($logic_or), ID($_AND_), ID($_OR_), ID($_XOR_))) {
-			hashlib::commutative_hash comm;
-			comm.eat((cell->getPort(ID::A)));
-			comm.eat((cell->getPort(ID::B)));
-			h = comm.hash_into(h);
-		} else if (cell->type.in(ID($reduce_xor), ID($reduce_xnor))) {
-			SigSpec a = (cell->getPort(ID::A));
-			a.sort();
-			h = a.hash_into(h);
-		} else if (cell->type.in(ID($reduce_and), ID($reduce_or), ID($reduce_bool))) {
-			SigSpec a = (cell->getPort(ID::A));
-			a.sort_and_unify();
-			h = a.hash_into(h);
-		} else if (cell->type == ID($pmux)) {
-			SigSpec sig_s = (cell->getPort(ID::S));
-			SigSpec sig_b = (cell->getPort(ID::B));
-			h = hash_pmux_in(sig_s, sig_b, h);
-			h = (cell->getPort(ID::A)).hash_into(h);
-		} else {
-			hashlib::commutative_hash comm;
-			for (const auto& [port, sig] : cell->connections()) {
-				if (cell->output(port))
-					continue;
-				comm.eat(hash_pair(port, (sig)));
-			}
-			h = comm.hash_into(h);
-			if (cell->is_builtin_ff())
-				h = initvals(cell->getPort(ID::Q)).hash_into(h);
-		}
-		return h;
-	}
-
-	static Hasher hash_cell_parameters(const RTLIL::Cell *cell, Hasher h)
-	{
-		hashlib::commutative_hash comm;
-		for (const auto& param : cell->parameters) {
-			comm.eat(param);
-		}
-		return comm.hash_into(h);
-	}
-
-	Hasher hash_cell_function(const RTLIL::Cell *cell, Hasher h) const
-	{
-		h.eat(cell->type);
-		h = hash_cell_inputs(cell, h);
-		h = hash_cell_parameters(cell, h);
-		return h;
-	}
-
-	bool compare_cell_parameters_and_connections(const RTLIL::Cell *cell1, const RTLIL::Cell *cell2) const
-	{
-		if (cell1 == cell2) return true;
-		if (cell1->type != cell2->type) return false;
-
-		if (cell1->parameters != cell2->parameters)
-			return false;
-
-		if (cell1->connections_.size() != cell2->connections_.size())
-			return false;
-		for (const auto &it : cell1->connections_)
-			if (!cell2->connections_.count(it.first))
-				return false;
-
-		decltype(Cell::connections_) conn1, conn2;
-		conn1.reserve(cell1->connections_.size());
-		conn2.reserve(cell1->connections_.size());
-
-		for (const auto &it : cell1->connections_) {
-			if (cell1->output(it.first)) {
-				if (it.first == ID::Q && cell1->is_builtin_ff()) {
-					// For the 'Q' output of state elements,
-					//   use the (* init *) attribute value
-					conn1[it.first] = initvals(it.second);
-					conn2[it.first] = initvals(cell2->getPort(it.first));
-				}
-				else {
-					conn1[it.first] = RTLIL::SigSpec();
-					conn2[it.first] = RTLIL::SigSpec();
-				}
-			}
-			else {
-				conn1[it.first] = (it.second);
-				conn2[it.first] = (cell2->getPort(it.first));
-			}
-		}
-
-		if (cell1->type.in(ID($and), ID($or), ID($xor), ID($xnor), ID($add), ID($mul),
-				ID($logic_and), ID($logic_or), ID($_AND_), ID($_OR_), ID($_XOR_))) {
-			if (conn1.at(ID::A) < conn1.at(ID::B)) {
-				std::swap(conn1[ID::A], conn1[ID::B]);
-			}
-			if (conn2.at(ID::A) < conn2.at(ID::B)) {
-				std::swap(conn2[ID::A], conn2[ID::B]);
-			}
-		} else
-		if (cell1->type.in(ID($reduce_xor), ID($reduce_xnor))) {
-			conn1[ID::A].sort();
-			conn2[ID::A].sort();
-		} else
-		if (cell1->type.in(ID($reduce_and), ID($reduce_or), ID($reduce_bool))) {
-			conn1[ID::A].sort_and_unify();
-			conn2[ID::A].sort_and_unify();
-		} else
-		if (cell1->type == ID($pmux)) {
-			sort_pmux_conn(conn1);
-			sort_pmux_conn(conn2);
-		}
-
-		return conn1 == conn2;
-	}
-
-	bool has_dont_care_initval(const RTLIL::Cell *cell)
-	{
-		if (!cell->is_builtin_ff())
-			return false;
-
-		return !initvals(cell->getPort(ID::Q)).is_fully_def();
-	}
+	CellHasher hasher;
 
 	OptMergeIncWorker(RTLIL::Design *design, RTLIL::Module *module, bool mode_nomux, bool mode_share_all, bool mode_keepdc) :
-		design(design), module(module), mode_share_all(mode_share_all)
+		design(design), module(module), mode_share_all(mode_share_all),
+		hasher(assign_map, initvals, /*apply_sigmap=*/false)
 	{
 		total_count = 0;
 		ct.setup_internals();
@@ -235,28 +78,6 @@ struct OptMergeIncWorker
 
 		initvals.set(&assign_map, module);
 
-
-		// We keep a set of known cells. They're hashed with our hash_cell_function
-		// and compared with our compare_cell_parameters_and_connections.
-		// Both need to capture OptMergeIncWorker to access initvals
-		struct CellPtrHash {
-			const OptMergeIncWorker& worker;
-			CellPtrHash(const OptMergeIncWorker& w) : worker(w) {}
-			std::size_t operator()(const Cell* c) const {
-				return (std::size_t)worker.hash_cell_function(c, Hasher()).yield();
-			}
-		};
-		struct CellPtrEqual {
-			const OptMergeIncWorker& worker;
-			CellPtrEqual(const OptMergeIncWorker& w) : worker(w) {}
-			bool operator()(const Cell* lhs, const Cell* rhs) const {
-				return worker.compare_cell_parameters_and_connections(lhs, rhs);
-			}
-		};
-		// std::unordered_set<
-		//     RTLIL::Cell*,
-		//     CellPtrHash,
-		//     CellPtrEqual> known_cells (0, CellPtrHash(*this), CellPtrEqual(*this));
 
 		dict<Cell *, uint32_t> hashes;
 		dict<uint32_t, Cell *> first_with_hash;
@@ -337,7 +158,7 @@ struct OptMergeIncWorker
 					}
 					if (cell->type == ID($scopeinfo))
 						continue;
-					if (mode_keepdc && has_dont_care_initval(cell))
+					if (mode_keepdc && hasher.has_dont_care_initval(cell))
 						continue;
 					if (!cell->known())
 						continue;
@@ -366,7 +187,7 @@ struct OptMergeIncWorker
 					}
 					if (cell->type == ID($scopeinfo))
 						continue;
-					if (mode_keepdc && has_dont_care_initval(cell))
+					if (mode_keepdc && hasher.has_dont_care_initval(cell))
 						continue;
 					if (!cell->known())
 						continue;
@@ -391,7 +212,7 @@ struct OptMergeIncWorker
 			pool<int32_t> buckets;
 
 			for (auto cell : cells) {
-				uint32_t hash = hash_cell_function(cell, Hasher()).yield();
+				uint32_t hash = hasher.hash_cell_function(cell, Hasher()).yield();
 				if (remember_cell(cell, hash))
 					buckets.emplace(hash);
 			}
@@ -412,7 +233,7 @@ struct OptMergeIncWorker
 						if (removed.count(cell))
 							break;
 
-						if (!compare_cell_parameters_and_connections(cell, other_cell))
+						if (!hasher.compare_cell_parameters_and_connections(cell, other_cell))
 							continue;
 
 						if (cell->has_keep_attr()) {
